cities variables stored separately. map syntax refined
