+++ src/Cities.jsx
@@ -1,0 +1,14 @@
+const cities = [
+  {
+    name: 'San Francisco',
+    latitude: 37.7853,
+    longitude: -122.41669,
+  },
+  {
+    name: 'Toronto',
+    latitude: 43.761539,
+    longitude: -79.411079,
+  },
+];
+
+export default cities;
